--- src/core/test/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  UnitTest: require("./UnitTest"),
-  IntegrationTest: require("./IntegrationTest")
-};
